(PC-22773)[BSR] refactor: full mail icon
