+++ pro/src/components/LegalInfos/LegalInfos.scss
@@ -12,9 +12,8 @@
   }
 
   .icon-legal-infos {
-    fill: colors.$primary;
     width: rem.torem(16px);
-    vertical-align: top;
+    vertical-align: text-bottom;
     margin-right: rem.torem(4px);
   }
 }
