+++ pro/src/icons/index.ts
@@ -22,7 +22,6 @@ export { ReactComponent as InfoPhoneIcon } from 'icons/info-phone.svg'
 export { ReactComponent as InfoWrongIcon } from 'icons/ico-info-wrong.svg'
 export { ReactComponent as KeyIcon } from 'icons/ico-key.svg'
 export { ReactComponent as LogoPassCultureDarkIcon } from 'icons/logo-pass-culture-dark.svg'
-export { ReactComponent as MailIcon } from 'icons/ico-mail.svg'
 export { ReactComponent as MailOutlineIcon } from 'icons/ico-mail-outline.svg'
 export { ReactComponent as MoreCircleIcon } from 'icons/ico-more-circle.svg'
 export { ReactComponent as NotificationErrorIcon } from 'icons/ico-notification-error.svg'
